+++ CHANGELOG.md
@@ -11,6 +11,8 @@ This component adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.h
 
 ### Changed
 
+- MongoDB instrumentation is updated to comply with v1.39.0 Semantic Convention
+
 #### Dependency updates
 
 ### Deprecated